pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "opsagent"
version = "0.2.0"
description = "SaaS de diagnóstico operativo con agentes IA para PyMEs industriales"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "langchain-anthropic>=0.3",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "fastapi>=0.110",
    "python-multipart>=0.0.9",
    "uvicorn>=0.29",
    "streamlit>=1.35",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "PyJWT>=2.8",
    "supabase>=2.0,<2.12",
    "reportlab>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests que requieren ANTHROPIC_API_KEY (excluidos por defecto)",
]
addopts = "-m 'not integration'"
